requirements.txt
streamlit==1.50.0
openai==2.15.0
python-dotenv==1.2.1
pydantic==2.12.5
supabase==2.27.2
postgrest==2.27.2
httpx==0.28.1
requests==2.32.5
tavily==1.1.0
pillow==11.3.0
pandas==2.3.3
numpy==2.0.2
